requirements.txt
Pillow>=9.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0